# openapi/app.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lattis application API",
    "version": "0.1.0-content-draft"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4100"
    }
  ],
  "paths": {
...
    "/api/videos/{id}/file": {
      "put": { "summary": "Stream an MP4 upload into private storage", "x-lattis-permission": "video.manage:video", "requestBody": { "required": true, "content": { "application/octet-stream": { "schema": { "type": "string", "format": "binary" } } } }, "responses": { "201": { "description": "Video uploaded" }, "400": { "description": "Hash or format mismatch" }, "401": { "description": "Unauthenticated" }, "403": { "description": "Forbidden" }, "409": { "description": "Upload conflict" }, "413": { "description": "Too large" } } }
    },
    "/api/videos/{id}/playback": {
      "post": { "summary": "Create a player-neutral attributed playback session with HLS/DASH CMAF sources when packaged", "x-lattis-permission": "video.play:video", "responses": { "200": { "description": "Playback descriptor v1 with sources, protection and watermark policy" }, "401": { "description": "Unauthenticated" }, "403": { "description": "Forbidden" }, "501": { "description": "Required DRM or forensic watermark is unavailable" } } }
    },
...
    "/api/videos/{id}/stream": {
      "get": { "summary": "Stream an authorized playback with byte ranges", "x-lattis-permission": "video.play:video", "responses": { "200": { "description": "Full MP4" }, "206": { "description": "MP4 byte range" }, "401": { "description": "Unauthenticated" }, "403": { "description": "Forbidden" }, "416": { "description": "Invalid range" } } }
    },
...
  },
...
}